BUG: Fix incorrect read of newer VTK files
VtkIO can read the .vtk file version 4.2.
In VTK 9, a new version of the .vtk file format, 5.1, was introduced.
This fix internally converts all files back to the 4.2 version so the
reading code is still valid.

The symptom was segfaults in some of the downstream modules that use
this library because the indices read were no longer correct or within
bounds.

In the future, this should probably be updated to not have its own file
reader, but rather to use VTK's, and to convert from vtkPolyData, or
use vtkPolyData directly.